+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Activity rows show a file, folder, or link icon beside the name.
+
 ## [0.3.8] - 2026-09-05
 
 ### Added